refactor(dry/class_manager): restructure class mgmt
And clean up the code.
