+++ dcp/api/__init__.py
@@ -1,9 +1,9 @@
 from .compute_for import compute_for
 from .job import Job
 
-inheritance_hooks = {
+sub_classes = {
     'Job': Job,
 }
 
-__all__ = ['compute_for', 'inheritance_hooks']
+__all__ = ['compute_for', 'sub_classes']
