+++ dcp/dry/__init__.py
@@ -1,6 +1,5 @@
-from .bfclass import make_dcp_class, wrap_js_obj
 from . import class_manager
-from .aio import asyncify, blockify
+from . import aio
 
-__all__ = ['make_dcp_class', 'wrap_js_obj', class_manager]
+__all__ = ['class_manager', 'aio']
